Only configure whitebox test suite if the a test src folder exists
This is mainly for the default 'test' source set, which is always
registered. If it is not used, the whitebox configuration will cause
Gradle to think that there are supposed to be tests, because the
main classes become part of the "test classes" set. This will cause
Gradle to run the task without tests and throw an error
(in recent Gradle versions).